move LinkedAgentCount component to top-level components and adjust output
