+++ x-pack/plugins/fleet/public/applications/fleet/components/linked_agent_count.tsx
@@ -5,31 +5,26 @@
  */
 
 import React, { memo } from 'react';
-import { FormattedMessage } from '@kbn/i18n/react';
 import { EuiLink } from '@elastic/eui';
-import { useLink } from '../../../hooks';
-import { AGENT_SAVED_OBJECT_TYPE } from '../../../constants';
+import { useLink } from '../hooks';
+import { AGENT_SAVED_OBJECT_TYPE } from '../constants';
 
+/**
+ * Displays the provided `count` number as a link to the Agents list if it is greater than zero
+ */
 export const LinkedAgentCount = memo<{ count: number; agentPolicyId: string }>(
   ({ count, agentPolicyId }) => {
     const { getHref } = useLink();
-    const displayValue = (
-      <FormattedMessage
-        id="xpack.fleet.agentPolicy.linkedAgentCountText"
-        defaultMessage="{count, plural, one {# agent} other {# agents}}"
-        values={{ count }}
-      />
-    );
     return count > 0 ? (
       <EuiLink
         href={getHref('fleet_agent_list', {
           kuery: `${AGENT_SAVED_OBJECT_TYPE}.policy_id : ${agentPolicyId}`,
         })}
       >
-        {displayValue}
+        {count}
       </EuiLink>
     ) : (
-      displayValue
+      <>count</>
     );
   }
 );
