+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/components/index.ts
@@ -8,7 +8,7 @@ export { AgentPolicyCopyProvider } from './agent_policy_copy_provider';
 export { AgentPolicyDeleteProvider } from './agent_policy_delete_provider';
 export { PackagePolicyDeleteProvider } from './package_policy_delete_provider';
 export { AgentPolicyYamlFlyout } from './agent_policy_yaml_flyout';
-export { LinkedAgentCount } from './linked_agent_count';
+export { LinkedAgentCount } from '../../../components/linked_agent_count';
 export { ConfirmDeployAgentPolicyModal } from './confirm_deploy_modal';
 export { DangerEuiContextMenuItem } from './danger_eui_context_menu_item';
 export { AgentPolicyActionMenu } from './actions_menu';
